requirements.txt
google-generativeai
pydantic
python-dotenv
pytest
streamlit
duckduckgo_search
